Move the CUDA end-to-end example to FFI examples workflow + hosted runner.
